+++ core/src/apps/management/recovery_device/__init__.py
@@ -23,9 +23,9 @@ async def recovery_device(ctx: wire.Context, msg: RecoveryDevice) -> Success:
     _check_state(msg)
 
     if not msg.dry_run:
-        title = "Wallet recovery"
+        title = "Recovery mode"
         text = Text(title, ui.ICON_RECOVERY)
-        text.normal("Do you really want to", "recover the wallet?", "")
+        text.normal("Do you really want to", "recover a wallet?", "")
     else:
         title = "Seed check"
         text = Text(title, ui.ICON_RECOVERY)
@@ -48,19 +48,21 @@ async def recovery_device(ctx: wire.Context, msg: RecoveryDevice) -> Success:
         newpin = await request_pin_confirm(ctx, allow_cancel=False)
         config.change_pin(pin_to_int(""), pin_to_int(newpin))
 
-    storage.device.set_u2f_counter(msg.u2f_counter)
+    if msg.u2f_counter:
+        storage.device.set_u2f_counter(msg.u2f_counter)
     storage.device.load_settings(
         label=msg.label, use_passphrase=msg.passphrase_protection
     )
     storage.recovery.set_in_progress(True)
-    storage.recovery.set_dry_run(msg.dry_run)
+    if msg.dry_run:
+        storage.recovery.set_dry_run(msg.dry_run)
 
     result = await recovery_process(ctx)
 
     return result
 
 
-def _check_state(msg: RecoveryDevice):
+def _check_state(msg: RecoveryDevice) -> None:
     if not msg.dry_run and storage.is_initialized():
         raise wire.UnexpectedMessage("Already initialized")
     if msg.dry_run and not storage.is_initialized():