+++ core/src/apps/management/recovery_device/homescreen.py
@@ -63,7 +63,7 @@ async def _finish_recovery_dry_run(
     ctx: wire.Context, secret: bytes, mnemonic_type: int
 ) -> Success:
     digest_input = sha256(secret).digest()
-    stored, _ = mnemonic.get()
+    stored = mnemonic.get_secret()
     digest_stored = sha256(stored).digest()
     result = utils.consteq(digest_stored, digest_input)
 
@@ -78,7 +78,7 @@ async def _finish_recovery_dry_run(
             == storage.recovery.get_slip39_iteration_exponent()
         )
 
-    await layout.show_dry_run_result(ctx, result)
+    await layout.show_dry_run_result(ctx, result, mnemonic_type)
 
     storage.recovery.end_progress()
 
@@ -95,17 +95,20 @@ async def _finish_recovery(
         secret, mnemonic_type, needs_backup=False, no_backup=False
     )
     if mnemonic_type == mnemonic.TYPE_SLIP39:
-        storage.device.set_slip39_identifier(storage.recovery.get_slip39_identifier())
-        storage.device.set_slip39_iteration_exponent(
-            storage.recovery.get_slip39_iteration_exponent()
-        )
-    storage.recovery.end_progress()
+        identifier = storage.recovery.get_slip39_identifier()
+        exponent = storage.recovery.get_slip39_iteration_exponent()
+        if identifier is None or exponent is None:
+            # Identifier and exponent need to be stored in storage at this point
+            raise RuntimeError
+        storage.device.set_slip39_identifier(identifier)
+        storage.device.set_slip39_iteration_exponent(exponent)
 
     beam_nonce_seed = random.bytes(32)
     create_beam_master_nonce(beam_nonce_seed)
 
     await show_success(ctx, ("You have successfully", "recovered your wallet."))
 
+    storage.recovery.end_progress()
     return Success(message="Device recovered")
 
 
@@ -122,7 +125,9 @@ async def _request_and_store_word_count(ctx: wire.Context, dry_run: bool) -> int
     return word_count
 
 
-async def _request_secret(ctx: wire.Context, word_count: int, mnemonic_type: int):
+async def _request_secret(
+    ctx: wire.Context, word_count: int, mnemonic_type: int
+) -> bytes:
     await _request_share_first_screen(ctx, word_count, mnemonic_type)
 
     secret = None
@@ -153,7 +158,7 @@ async def _request_secret(ctx: wire.Context, word_count: int, mnemonic_type: int
 
 async def _request_share_first_screen(
     ctx: wire.Context, word_count: int, mnemonic_type: int
-):
+) -> None:
     if mnemonic_type == mnemonic.TYPE_BIP39:
         content = layout.RecoveryHomescreen(
             "Enter recovery seed", "(%d words)" % word_count
@@ -172,9 +177,12 @@ async def _request_share_first_screen(
         raise RuntimeError
 
 
-async def _request_share_next_screen(ctx: wire.Context, mnemonic_type: int):
+async def _request_share_next_screen(ctx: wire.Context, mnemonic_type: int) -> None:
     if mnemonic_type == mnemonic.TYPE_SLIP39:
         remaining = storage.recovery.get_remaining()
+        if not remaining:
+            # 'remaining' should be stored at this point
+            raise RuntimeError
         if remaining == 1:
             text = "1 more share"
         else: