+++ core/src/apps/management/recovery_device/layout.py
@@ -7,9 +7,9 @@ from trezor.ui.info import InfoConfirm
 from trezor.ui.text import Text
 from trezor.ui.word_select import WordSelector
 
-from .bip39_keyboard import Bip39Keyboard
+from .keyboard_bip39 import Bip39Keyboard
+from .keyboard_slip39 import Slip39Keyboard
 from .recover import RecoveryAborted
-from .slip39_keyboard import Slip39Keyboard
 
 from apps.common import mnemonic, storage
 from apps.common.confirm import confirm, require_confirm
@@ -39,7 +39,7 @@ async def request_word_count(ctx: wire.Context, dry_run: bool) -> int:
     if dry_run:
         text = Text("Seed check", ui.ICON_RECOVERY)
     else:
-        text = Text("Wallet recovery", ui.ICON_RECOVERY)
+        text = Text("Recovery mode", ui.ICON_RECOVERY)
     text.normal("Number of words?")
 
     if __debug__:
@@ -86,25 +86,41 @@ async def request_mnemonic(
     return " ".join(words)
 
 
-async def show_dry_run_result(ctx: wire.Context, result: bool) -> None:
+async def show_dry_run_result(
+    ctx: wire.Context, result: bool, mnemonic_type: int
+) -> None:
     if result:
-        await show_success(
-            ctx,
-            (
+        if mnemonic_type == mnemonic.TYPE_SLIP39:
+            text = (
                 "The entered recovery",
-                "seed is valid and matches",
-                "the one in the device.",
-            ),
-        )
+                "shares are valid and",
+                "match what is currently",
+                "in the device.",
+            )
+        else:
+            text = (
+                "The entered recovery",
+                "seed is valid and",
+                "matches the one",
+                "in the device.",
+            )
+        await show_success(ctx, text, button="Continue")
     else:
-        await show_warning(
-            ctx,
-            (
+        if mnemonic_type == mnemonic.TYPE_SLIP39:
+            text = (
                 "The entered recovery",
-                "seed is valid but does not",
-                "match the one in the device.",
-            ),
-        )
+                "shares are valid but",
+                "do not match what is",
+                "currently in the device.",
+            )
+        else:
+            text = (
+                "The entered recovery",
+                "seed is valid but does",
+                "not match the one",
+                "in the device.",
+            )
+        await show_warning(ctx, text, button="Continue")
 
 
 async def show_dry_run_different_type(ctx: wire.Context) -> None:
@@ -134,45 +150,33 @@ async def show_keyboard_info(ctx: wire.Context) -> None:
         await ctx.wait(info)
 
 
-async def show_invalid_mnemonic(ctx, mnemonic_type: int):
+async def show_invalid_mnemonic(ctx: wire.Context, mnemonic_type: int) -> None:
     if mnemonic_type == mnemonic.TYPE_SLIP39:
-        await show_warning(
-            ctx,
-            ("You have entered", "a recovery share", "that is not valid."),
-            button="Try again",
-        )
+        await show_warning(ctx, ("You have entered", "an invalid recovery", "share."))
     else:
-        await show_warning(
-            ctx,
-            ("You have entered", "a recovery seed", "that is not valid."),
-            button="Try again",
-        )
+        await show_warning(ctx, ("You have entered", "an invalid recovery", "seed."))
 
 
-async def show_share_already_added(ctx):
-    return await show_warning(
-        ctx,
-        ("Share already entered,", "please enter", "a different share."),
-        button="Try again",
+async def show_share_already_added(ctx: wire.Context) -> None:
+    await show_warning(
+        ctx, ("Share already entered,", "please enter", "a different share.")
     )
 
 
-async def show_identifier_mismatch(ctx):
-    return await show_warning(
-        ctx,
-        ("You have entered", "a share from another", "Shamir Backup."),
-        button="Try again",
+async def show_identifier_mismatch(ctx: wire.Context) -> None:
+    await show_warning(
+        ctx, ("You have entered", "a share from another", "Shamir Backup.")
     )
 
 
-class RecoveryHomescreen(ui.Control):
+class RecoveryHomescreen(ui.Component):
     def __init__(self, text: str, subtext: str = None):
         self.text = text
         self.subtext = subtext
         self.dry_run = storage.recovery.is_dry_run()
         self.repaint = True
 
-    def on_render(self):
+    def on_render(self) -> None:
         if not self.repaint:
             return
 
@@ -204,8 +208,6 @@ async def homescreen_dialog(
     ctx: wire.Context, homepage: RecoveryHomescreen, button_label: str
 ) -> None:
     while True:
-        # make sure the homepage gets painted, even after cancelling the dialog
-        homepage.repaint = True
         continue_recovery = await confirm(
             ctx,
             homepage,