+++ core/src/trezor/messages/Features.py
@@ -5,8 +5,11 @@ import protobuf as p
 if __debug__:
     try:
         from typing import Dict, List, Optional
+        from typing_extensions import Literal  # noqa: F401
+        EnumTypeFeature = Literal[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
     except ImportError:
         Dict, List, Optional = None, None, None  # type: ignore
+        EnumTypeFeature = None  # type: ignore
 
 
 class Features(p.MessageType):
@@ -41,6 +44,8 @@ class Features(p.MessageType):
         fw_vendor_keys: bytes = None,
         unfinished_backup: bool = None,
         no_backup: bool = None,
+        recovery_mode: bool = None,
+        features: List[EnumTypeFeature] = None,
     ) -> None:
         self.vendor = vendor
         self.major_version = major_version
@@ -69,6 +74,8 @@ class Features(p.MessageType):
         self.fw_vendor_keys = fw_vendor_keys
         self.unfinished_backup = unfinished_backup
         self.no_backup = no_backup
+        self.recovery_mode = recovery_mode
+        self.features = features if features is not None else []
 
     @classmethod
     def get_fields(cls) -> Dict:
@@ -100,4 +107,6 @@ class Features(p.MessageType):
             26: ('fw_vendor_keys', p.BytesType, 0),
             27: ('unfinished_backup', p.BoolType, 0),
             28: ('no_backup', p.BoolType, 0),
+            29: ('recovery_mode', p.BoolType, 0),
+            30: ('features', p.EnumType("Feature", (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14)), p.FLAG_REPEATED),
         }